add Windows debug machine
Today we don't have any Windows machine in the CI pool. The machine
template has not changed since 2019-11-21, yet as of today when the
machine starts GCP proudly declares

> GCEMetadataScripts: No startup scripts to run.

despite the script being defined as `sysprep-specialize-script-ps1`, as
per the
[documentation](https://cloud.google.com/compute/docs/startupscript).
Also, it used to work and we haven't changed anything.

I'm not quite sure what's going on and how to investigate, but I think
at the very least we can try to unblock the team by having a set of
machines we initialize manually. This PR is meant to do that.)

This is the same changeset as a87749113915808b8238c6e5e0d77f6adc2dd143
and 16da7005326e29c44c15c55a2bc33c1d8575ffdb, except that it now
specifies 5 machines instead of just one.

CHANGELOG_BEGIN
CHANGELOG_END