Fix indexing of transactions referencing divulged contracts (#4442)
* Fix indexing of transactions referencing divulged contracts

When a transaction referenced a divulged contract that had no contract
metadata (e.g. it was only known to the participant via divulgence),
the indexer failed with InactiveDependencyError.

This fixes the issue by modifying lookupContractLet query to return
LetUnknown if the contract is known, but has no metadata.

No changelog entry to add as this bug did not affect any users as this
surfaced with privacy extensions. Canton not affected as they always
include divulged contracts, even if they already have been divulged.

CHANGELOG_BEGIN
CHANGELOG_END

* Failing test-case for lookupContractLet

* Address PR review

* Second pass addressing review