prepare playground for perf testing
The goal of this commit is to provide me with a playground to test out
various GCP machine configurations and assess what impact they have on
build performance. I need separate pools for that as changing the
machine type requires a machine restart, which would be disruptive to
any running job if I did that in the existing pools.

The first commit in this PR does the following:

- Deactivate Bazel caching.
- Create two new machine groups (of size 1) for Linux and Windows pools.
- Disable all jobs that are not building the project on Linux or
  Windows.
- Change the Linux and Windows jobs to use the new pools.

The plan is for me to manually change the machine types in the new pool
and rerun the build repeatedly. Once I believe I have the information I
need, the expected end result for this PR is only a change to the
machine type (or no change at all if the current machine size is already
optimal). All other changes will be undone and the additional pools will
be destroyed.

Note: as of the opening of this PR, the two new pools have not been
created yet at the Azure level (this is a separate, manual step not
managed by Terraform) and the Terraform files have not been applied.

CHANGELOG_BEGIN
CHANGELOG_END