TypeScript: Make contract ids type safe
Currently, contract ids are simply presented by strings. Thus, it is very
easy to accidentally mix up contract ids of different templates. This PR
is an attempt to provide more safety in this regard, namely that contract
ids of two templates with _structurally_ different types cannot be mixed
up. The implementation is slightly complicated since we need to be able
to deal with contract ids of unknown templates as well.

You're feedback on this PR is very welcome.

A simpler alternative would be to use the type `T` for the brand of
`ContractId<T>`. However, this would have the effect that `ContractId`
is covariant in its argument and we hence get a subtyping relationship
between `ContractId<T>` and `ContractId<U>` whenever `T` and `U` are
in a subtyping relationship. Since TypeScript is structurally typed, the
types for two different DAML templates might end up in a relationship
easily, particularly when you have templates whose data is very similar
since they are successive steps of a single workflow.

CHANGELOG_BEGIN
CHANGELOG_END