Remove hardcoded number of packages from Haskell ledger binding tests
This is one of the tests that keeps breaking every time we change the
number of packages `damlc build` outputs by default which is quite
annoying.

The actual change is trivial, we just read the number of packages from
the manifest. The diff is mostly just propagating this through everything.

changelog_begin
changelog_end