only run release if build succeeded
The default behaviour of an Azure job that has a dependency is to only
run if the dependency has succeeded. However, that default behaviour is
overridden if there is an exmplicit `condition` attribute.

This PR restores the expected behaviour that we only try to build a
release tarball if the actual build has succeeded.

CHANGELOG_BEGIN
CHANGELOG_END