New bazel rule -- `daml_build_test` -- to build a DAML project from its `daml.yaml` config, and setup a bazel test rule to validate the resulting .dar using `daml validate-dar`.
Dependencies on other DAML projects are declared with the `dar_dict` attribute of the build rule. This attribute also declares the names by which the `.dar` files are known in the client project, corresponding to the references in the `daml.yaml` config.

The new rule is used build & test the upgrade documentation example code.

changelog_begin
changelog_end